% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body}
\alias{body}
\title{Create a humap body plot}
\usage{
body(data, loc, side = NULL, type = "simple", proj = "neutral",
  body_halves = "separate", annotate = "freq", bridge = NULL,
  na_rm = FALSE, combine = NULL, controls = NULL)
}
\arguments{
\item{data}{tidy data frame (more under details)}

\item{loc}{column in data frame containing localisation codes for
observations. Character string.}

\item{side}{variable indication left and right side of observations. Values
must be "left" or "right". May also be a named *list* that defines the variable (=
data frame column) holding the variables, and which value(s) in that column
correspond(s) to left and right. Valid specifications are, i.a., \code{side = "xx"} and \code{side = list(var = "xx",
left = c("left", "l"), right = "r")}}

\item{type}{simple, (default), f(emale), m(male)}

\item{proj}{"front" or "back". Ignored if \code{type = "simple"}.}

\item{body_halves}{character string defining how to deal with body halves.
"separate" (default) discriminates the left half from the right; "join" merges
observations in, e.g., right and left side of the chest.}

\item{annotate}{\code{freq} (defaults) shows only absolute and relative
frequencies. \code{all} include region names. \code{none} omits labels
altogether. More elaborate argument may be given in \code{list} form, see
Details below.}

\item{bridge}{used to bridge between the user's data format and the native
format of humapr. Further explanations to follow}

\item{na_rm}{should missing data be removed? Defaults is \code{FALSE}.}

\item{combine}{a \emph{list} of vectors naming the regions to be combined and
mapped as one. E.g., \code{list(arm = c("shoulder", "arm", "elbow",
"wrist"))}}

\item{controls}{list of controls for fine-tuning. See details.}
}
\value{
A ggplot object with suitable layout settings for the purpose of
  \code{body}.
}
\description{
\code{body} creates a ggplot object, using its own set of "geoms", i.e.,
human body oulines on which your data are projected in a heatmap-like
fashion. \code{humapr} is an extension of ggplot2, and objects created with
the \code{humap} function can be modified with standard ggplot2
setting--e.g., calls to \code{theme()}.
}
\details{
If you don't supply a \code{body_halves} argument humapr will assume default
to \code{pool} mode.

\code{controls} is a list of more specific parameters to modify details in
the apperance of the humap. These controls are currently available:
\code{vert_adj} has "minimal" as default, but may also be set to "even" for
vertical even distribution, or "smart" to avoid crossing annovation lines.
\code{na_fill} specificies the fill of zero-count regions. \code{mid_include}
is a logical indicating whether midline observations should be included;
currently, if this is set to TRUE, a midline observation will count as half
an observations on each side. Default is FALSE.
}
