% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ais.R
\name{geom_ais}
\alias{geom_ais}
\title{geom_ais builds a body map using AIS coding directly}
\usage{
geom_ais(mapping = NULL, data = NULL, type = "simple",
  with_back = TRUE, body_halves = NULL, annotate = "freq",
  bridge_loc = NULL, bridge_side = NULL, combine = NULL,
  controls = NULL, na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{supply an \code{aes()} object, just like other \code{ggplot2} functions. \code{humapr::geom_*} can
take to aesthetics: \code{loc} is the name of the data frame column containing localisation codes for observations;
and \code{side} the column holding the laterality data (must be either "left" or "right, although see
\code{bridge_side}).}

\item{data}{data frame, like the type you'd feed into \code{ggplot()} when producing a
histogram. Is expected to be tidy, but can contain a summary statistic; in that
case, you must specify a \code{fill} aesthetic and set \code{stat = "identity"} in
the call to the \code{geom}.}

\item{type}{currently, only \code{"simple"} is available, but we're working on more
natural-looking and sex-specific maps as well.}

\item{with_back}{logical, should the plot include the back-side of the head, neck and trunk?}

\item{body_halves}{character string defining how to deal with body halves.
\code{"separate"} (default with side aesthetic) discriminates the left half from the
right; \code{"join"} (default without side aesthetic) merges observations in, e.g.,
right and left side of the chest.}

\item{annotate}{\code{"freq"} (defaults) shows only absolute and relative frequencies,
\code{"all"} includes region names, and \code{NA} omits labels altogether. See
Details for ways to fine-tune the apperance of annotations.}

\item{bridge_loc}{named \emph{list} specifying the bridge from your data format to the
native format of \code{geom_body}, e.g., \code{list(head = c("head", "face",
"scalp"))}. See Details.}

\item{bridge_side}{named \emph{list}, specyfing the bridge from your laterality values
for each of the native values of \code{geom_body}, e.g. \code{list(left = c("left",
"l"), right = c("right", "r"), mid = c("mid", "m"))}. The \code{mid} element is only
required if mid-line observation are included, see the \code{controls} argument.}

\item{combine}{named \emph{list} of vectors naming the regions to be combined and
mapped as one, e.g., \code{list(arm = c("shoulder", "arm", "elbow", "wrist"))}. See
Details for behaviour.}

\item{controls}{named \emph{list} of more specific parameters for fine-tuning the
appereance of the humap. Currently, these controls are available (more will follow
in the future): \itemize{ \item \code{na_fill}:  should be a string or function
specifying the fill of zero-count regions (default: "black") \item \code{label_pad}:
a numeric defining the padding between labels, in percent of map height (default:
\code{3.5}) \item \code{mid_include}: a logical defining whether to include mid-line
observations and split them equally between left and right (default: \code{FALSE})
\item \code{round_counts}: a logical indicating whether to round (up) half-counts in
annotation labels when splitting mid-line observations between left and right
(default: \code{FALSE}) }}

\item{na.rm}{logical indicating whether to remove missing data. Default is
\code{FALSE}.}

\item{show.legend}{passed on to the underlying machinery of the
\code{geom}.}

\item{inherit.aes}{passed on to the underlying machinery of the
\code{geom}.}

\item{...}{passed on to the underlying machinery of the
\code{geom}.}
}
\value{
A layer object to a \code{humap} object.
}
\description{
Use this function to draw a choropleth on a human body map. Use the arguments to choose exactly what kind of map you
want. If you don't supply a \code{body_halves} argument humapr will default to \code{pool} mode. \code{geom_ais} is a
variant of \code{\link{geom_body}}, so for see the details of that function for more information.
}
