% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{housekeeping}
\alias{housekeeping}
\title{Initial housekeeping in \code{humap} call}
\usage{
housekeeping(user, defs)
}
\arguments{
\item{user}{user-given argument; essentially useful stuff from \code{match.call()}}

\item{defs}{default argument values; essentially useful stuff from \code{formals()}}
}
\description{
Internal function. Sets up an environment with settings; tests argument values. I cannot think of any scenario in which a user would need to call this function
}
