% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humap.R
\name{humap}
\alias{humap}
\title{Create a humap, ready for plotting a geom}
\usage{
humap(data = NULL, mapping = NULL)
}
\arguments{
\item{data}{data frame, like the type you'd feed into \code{ggplot()} when producing a
histogram. Is expected to be tidy, but can contain a summary statistic; in that
case, you must specify a \code{fill} aesthetic and set \code{stat = "identity"} in
the call to the \code{geom}.}

\item{mapping}{\code{aes()} object, just like other \code{ggplot2} functions, with two
aesthetics: \code{loc} is the name of the data frame column containing localisation
codes for observations; and \code{side} the column holding the laterality data (must
be either "left" or "right, but see \code{bridge_side}). It is possible to supply a
\code{fill} argument to use pre-computed summary statistics (see below).}
}
\value{
A ggplot object with suitable layout settings for the purpose of
  \code{humap_body}.
}
\description{
\code{humap} sets up a \code{ggplot} object with \code{theme} settings appropriate for
adding \code{geoms} from the \pkg{humapr} package. A \code{ggplot} object, it's
compatible with standard \pkg{ggplot2} function, e.g., \code{theme()} and
\code{facet_wrap()}. Doesn't really do anything on its own.
}
