% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humapr.R
\docType{package}
\name{humapr}
\alias{humapr}
\alias{humapr-package}
\title{humapr: A package for visualising topographic human data in R}
\description{
The humapr package is a simple way to present topographic data from human in an appealing and intuitive way. It's still in its early development with a few \code{geoms}, but more will follow. Hopefully, we'll get feedback from the community on what geoms would be useful, and we'll try to keep up with the requests.
}
\details{
Please, make sure to report bugs and weird/unexpected outputs on our GitHub page, so we can improve the stability of the package, and make it robust to real-life use.
}
