% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_label.R
\name{make_label}
\alias{make_label}
\title{Create textGrob for humap annotation}
\usage{
make_label(id, data, local_coords)
}
\arguments{
\item{id}{internal id (e.g., right_abdomen) of label for which to create
textGrob}

\item{data}{the original data sent into \code{humap} by \code{ggplot2}}

\item{local_coords}{data fram holding coordinates for all annotations;
\code{id} is used to pick appropriate coordinates}
}
\description{
Internal function called within \code{geom_humap}. I cannot think of any
scenario in which a user would need to call this function.
}
