% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_humap.R
\name{geom_body}
\alias{geom_body}
\title{geom_body builds a normal body map}
\usage{
geom_body(mapping = NULL, data = NULL, type = "simple",
  proj = "neutral", body_halves = "separate", annotate = "freq",
  bridge_loc = NULL, bridge_side = NULL, combine = NULL,
  controls = NULL, na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{supply an \code{aes()} object, just like other \code{ggplot2}
functions. \code{humapr::geom_*} can take to aesthetics: \code{loc} is the
name of the data frame column containing localisation codes for
observations; and \code{side} the column holding the laterality data (must
be either "left" or "right, although see \code{bridge_side}).}

\item{data}{tidy data frame, like the type you'd feed into ggplot() when
producing a histogram}

\item{type}{currently, only \code{"simple"} is available, but we're working
on more natural-looking and gender-specific maps as well.}

\item{proj}{\code{"front"} or \code{"back"}. Ignored if \code{type =
"simple"}, so currently not in use.}

\item{body_halves}{character string defining how to deal with body halves.
\code{"separate"} (default) discriminates the left half from the right;
\code{"join"} merges observations in, e.g., right and left side of the
chest.}

\item{annotate}{\code{"freq"} (defaults) shows only absolute and relative
frequencies, \code{"all"} includes region names, and \code{NA} omits labels
altogether. See Details for ways to fine-tune the apperance of annotations.}

\item{bridge_loc}{named \emph{list} specifying the bridge from your data
format to the native format of \code{geom_body}, e.g., \code{list(head =
c("head", "face", "scalp"))}. See Details.}

\item{bridge_side}{names \emph{list}, specyfing the bridge from your
laterality values for each of the native values of \code{geom_body}, e.g.
\code{list(left = c("left", "l"), right = c("right", "r"), mid = c("mid",
"m"))}. The \code{mid} element is only required if mid-line observation are
included, see the \code{controls} argument.}

\item{combine}{named \emph{list} of vectors naming the regions to be combined and
mapped as one, e.g., \code{list(arm = c("shoulder", "arm", "elbow",
"wrist"))}. See Details.}

\item{controls}{named \emph{list} of more specific parameters for fine-tuning
the appereance of the humap. Currently, these controls are available (more
will follow in the fuuture): \code{na_fill} should be a string or function
specifying the fill of zero-count regions (default: "black"),
\code{label_pad} a numeric defining the padding between labels, in percent
of map height (default: \code{3.5}), \code{mid_include} a logical defining
whether to include mid-line observations and split them equally between
left and right (default: \code{FALSE}), and \code{round_counts} a logical
indicating whether to round (up) half-counts in annotation labels when
splitting mid-line observations between left and right (default:
\code{FALSE}).}

\item{na.rm}{logical indicating whether to remove missing data. Default is
\code{FALSE}.}

\item{show.legend, inherit.aes, ...}{like all other \code{geoms}.}
}
\value{
A layer object to a \code{humap} object.
}
\description{
Use this function to draw a choropleth on a human body map. Use the arguments
to choose exactly what kind of map you want. If you don't supply a
\code{body_halves} argument humapr will default to \code{pool} mode.
}
\details{
The \code{annotate} argument may be a named \code{list} with any combination
of the following four elements. The \code{details} element should be a
character string, and may be any of the valid values of the simple argument
(\code{"freq"}, \code{"all"}, \code{NA}). The remaining three elements
\code{gp}, \code{gp_text} and \code{gp_lines} must be \code{gpar} objects
(see \code{?grid::gpar}). If only \code{gp} is specified, it is applied to
both annotation text and lines. \code{gp_text} and \code{gp_lines} are
applied to annotation text and lines individually.

The \code{bridge} argument allows you to use your own localisation values,
and bridge them to those of \code{geom_body}. The idea is simple: you supply
a named \code{list}, each element of which corresponds one region in the
\code{geom_body} map; the name of the element should be the name of the
region, and the element should be a character vector of values in your data
that correspond to that region. See vignette [add reference to vignette here]
for examples.

If you want to combine several regions and map them as one, supply a named
\emph{list} in \code{combine}, following this logic: the name of each element
will be printed as the annotation text (if you so desire), and the element
must be a character vector specifying the names of \code{geom_body} regions
to map as one. Underscores in the list element names will be converted to
wide spaces. Make sure to \emph{not} use a name for a merged region that is
already used for another region; e.g., this is not allowed, as \code{hand}
already the name of another region: \code{list(hand = c("shoulder", "arm",
"elbow", "wrist"))}. You may, however, use a name of a region inside the
group, so this would be allowed: \code{list(hand = c("wrist", "hand"))}.
}
