% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{housekeeping}
\alias{housekeeping}
\title{Initial housekeeping in \code{humap} call}
\usage{
housekeeping(user, defs)
}
\arguments{
\item{user}{argument values from user, essentially useful stuff from match.call()}

\item{defs}{default values, essentially useful stuff from formals().}
}
\description{
Internal function. I cannot think of any scenario in which a user would need to call this function
}
