% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humap.R
\name{humap}
\alias{humap}
\title{Create a humap plot}
\usage{
humap(data, loc.var, lr.var = NULL, type = "body", gender = "neutral",
  proj = "simple", half = "both", annotate = "freq", anno_gp = NULL,
  bridge = NULL, na_rm = FALSE, combine = NULL, controls = NULL)
}
\arguments{
\item{data}{tidy data frame (more under details)}

\item{loc.var}{column in data frame containing localisation codes for
observations. Character string.}

\item{lr.var}{variable indication left and right side of observations. Values
must be "left" or "right". May also be a list that defines the variable (=
data frame column) holding the variables, and which value(s) in that column
correspond(s) to left and right. So, lr.var = "xx" or lr.var = list(var =
"xx", left = c("left", "l"), right = "r")}

\item{type}{body}

\item{gender}{n(eutral) (default), f(emale), m(male)}

\item{proj}{character string defining the projected front, back}

\item{half}{character string defining how to deal with body halves. "both"
discriminates the left half from the right; "mirror" merges observations
in, e.g., right and left side of the chest; "left"/"right" yields only the
indicated half and ignores observations on the other half.}

\item{annotate}{\code{freq} (defaults) shows only absolute and relative
frequencies. \code{all} include region names. \code{none} omits labels
altogether.}

\item{anno_gp}{grid::gpar object. Default is fontsize = 9 and col = "black".}

\item{bridge}{used to bridge between the user's data format and the native
format of humapr. Further explanations to follow}

\item{na_rm}{should missing data be removed? Defaults is \code{FALSE}.}

\item{combine}{a \emph{list} of vectors naming the regions to be combined and
mapped as one. E.g., list(arm = c("shoulder", "arm", "elbow", "wrist"))}

\item{controls}{list of controls for fine-tuning. See details.}
}
\value{
A ggplot object with layout settings suitable for the kinds of geoms
  used in \code{humap}.
}
\description{
\code{humap} creates a ggplot object, using its own set of "geoms", i.e.,
human body oulines on which your data are projected in a heatmap-like
fashion. \code{humapr} is an extension of ggplot2, and objects created with
the \code{humap} function can be modified with standard ggplot2
setting--e.g., calls to \code{theme()}.
}
\details{
If you don't supply a `side` argument humapr will assume default to mirroring
mode. If `side` argument is supplied, it will default to left/right
discrimate mode and show both body halves

If proj = "both" you cannot map right/left-discrimate data.

\code{controls} is a list of more specific parameters to modify details in
the apperance of the humap. These controls are currently available:
\code{vert_adj} has "minimal" as default, but may also be set to "even" for
vertical even distribution, or "smart" to avoid crossing annovation lines.
\code{na_fill} specificies the fill of zero-count regions. \code{mid_include}
is a logical indicating whether midline observations should be included;
currently, if this is set to TRUE, a midline observation will count as half
an observations on each side. Default is FALSE.
}
