% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humap.R
\name{humap}
\alias{humap}
\title{Create a humap, ready for plotting a geom}
\usage{
humap()
}
\value{
A ggplot object with suitable layout settings for the purpose of
  \code{humap_body}.
}
\description{
\code{humap} sets up a \code{ggplot} object with \code{theme} settings appropriate for adding \code{geoms} from the \pkg{humapr} package.
A \code{ggplot} object, it's compatible with standard \pkg{ggplot2} function, e.g.,
\code{theme()} and \code{facet_wrap()}. Doesn't really do anything on its own.
}
