% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humap_body.R
\name{humap}
\alias{humap}
\title{Create a humap, ready for plotting a geom}
\usage{
humap()
}
\value{
A ggplot object with suitable layout settings for the purpose of
  \code{hump_body}.
}
\description{
\code{humap_body} sets up a \code{ggplot} object, ready for drawing a map,
and compatible with with standard \code{ggplot2} function, e.g.,
\code{theme()} and \code{facet_wrap()}. Doesn't really do  much on its own.
}
